refactor(cli): split src/cli.js into command-registration modules (audit rec #4)
Audit recommendation #4: src/cli.js had grown to 699 LOC with 26
commands wired in a single file. Refactored into a thin entry point
plus six register modules, all under the 250-LOC target the audit
implied.

## Layout

  src/cli.js                          113 LOC  (was 699)
  src/cli/_shared.js                   30 LOC
  src/cli/register-pipeline.js        212 LOC  init/config/run/code/review/
                                                scan/doctor/status/report
  src/cli/register-plan.js            152 LOC  plan + 8 sub-commands
                                                (incl. REASONS Canvas autodetect)
  src/cli/register-meta.js            204 LOC  discover/triage/researcher/
                                                architect/audit/resume/update/
                                                board/undo/sync/clean
  src/cli/register-roles-skills.js     38 LOC  skills/roles/agents
  src/cli/register-sonar.js            28 LOC  sonar parent + 5 sub-commands

Each `register-*.js` exports one function `register*(program, { pkgVersion })`
that takes the Commander instance and adds its commands. cli.js
becomes:

  const program = new Command();
  // program-level options & version
  registerPipeline(program, { pkgVersion });
  registerPlan(program, { pkgVersion });
  registerMeta(program, { pkgVersion });
  registerRolesSkills(program, { pkgVersion });
  registerSonar(program, { pkgVersion });
  program.parse();

## Layout deviation

`triage` lives in `register-meta.js`, not pipeline. Reason: triage is
a single-role analysis command in the same family as discover /
researcher / architect / audit — none run the full agent loop. Keeping
all five together is more cohesive than splitting `triage` off.

## eslint.config.js

Added `src/cli/**/*.js` to the `no-console: off` allow-list. CLI
modules legitimately print user-facing output, same rationale as
src/cli.js itself.

Verified:
- node bin/kj.js --help                 → all 26 commands listed
- node bin/kj.js plan --help            → all 8 sub-commands
- node bin/kj.js {run,audit,sonar} --help → full flag sets intact
- npx eslint src/                       → 0 errors
- npx vitest run                        → 4192/4192 passing
